fix: fix ratings

+++ src/main/java/com/marcos/microservice/bookmanager/web/dto/AddReviewRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class AddReviewRequest {
-    int rating;
+    float rating;
     String review;
 }
